# api/openapi/current.json
{
  "components": {
    "schemas": {
      "AdaptiveQuizLoop": {
        "properties": {
          "captionArtifactId": {
            "pattern": "^[1-9]\\d*$",
            "type": "string"
          },
          "captionGeneration": {
            "minimum": 1,
            "type": "integer"
          },
          "failureCode": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "questions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "state": {
            "enum": [
              "generating",
              "ready",
              "evaluated",
              "failed",
              "stale"
            ],
            "type": "string"
          },
          "studyContextId": {
            "pattern": "^[1-9]\\d*$",
            "type": "string"
          },
          "watchedRange": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "id",
          "studyContextId",
          "state",
          "watchedRange",
          "captionArtifactId",
          "captionGeneration",
          "questions",
          "failureCode"
        ],
        "type": "object"
      },
      "AdaptiveQuizSubmission": {
        "properties": {
          "attempt": {
            "additionalProperties": true,
            "type": "object"
          },
          "reviewProposal": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "state": {
            "enum": [
              "evaluated"
            ],
            "type": "string"
          }
        },
        "required": [
          "state",
          "attempt",
          "reviewProposal"
        ],
        "type": "object"
      },
...
      "AgentRunResponse": {
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AgentRunAttemptResponse"
            },
            "type": "array"
          },
          "budgets": {
            "$ref": "#/components/schemas/AgentBudgetsResponse"
          },
          "cancellationRequestedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "courseId": {
            "nullable": true,
            "type": "integer"
          },
          "failureCode": {
            "nullable": true,
            "type": "string"
          },
          "finishedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "input": {
            "additionalProperties": true,
            "type": "object"
          },
          "ownerId": {
            "type": "integer"
          },
          "proposedSteps": {
            "items": {
              "$ref": "#/components/schemas/ProposedCourseStepResponse"
            },
            "type": "array"
          },
          "queuedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "awaiting_approval",
              "approved",
              "completed",
              "failed",
              "cancelled"
            ],
            "type": "string"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/AgentRunTransitionResponse"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/AgentUsageResponse"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "ownerId",
          "courseId",
          "state",
          "version",
          "input",
          "budgets",
          "usage",
          "queuedAt",
          "startedAt",
          "finishedAt",
          "updatedAt",
          "cancellationRequestedAt",
          "failureCode",
          "attempts",
          "transitions",
          "proposedSteps"
        ],
        "type": "object"
      },
...
      "ApproveLearningProposalDto": {
        "properties": {
          "courseId": {
            "minimum": 1,
            "type": "number"
          },
          "expectedCourseVersion": {
            "minimum": 1,
            "type": "number"
          },
          "proposalId": {
            "format": "uuid",
            "type": "string"
          },
          "targetKind": {
            "enum": [
              "existing_course",
              "new_private_course"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "proposalId",
          "targetKind"
        ],
        "type": "object"
      },
...
      "CreateAgentRunDto": {
        "properties": {
          "budgets": {
            "$ref": "#/components/schemas/AgentRunBudgetsDto"
          },
          "objective": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "requestedStepCount": {
            "maximum": 6,
            "minimum": 3,
            "type": "number"
          },
          "studyContextId": {
            "pattern": "^[1-9]\\d*$",
            "type": "string"
          },
          "watchedRanges": {
            "items": {
              "$ref": "#/components/schemas/AgentRunWatchedRangeDto"
            },
            "maxItems": 20,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "objective"
        ],
        "type": "object"
      },
...
      "ExpectedVersionDto": {
        "properties": {
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "LearningProgressResponse": {
        "properties": {
          "bestQuizScore": {
            "nullable": true,
            "type": "number"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "courseStepId": {
            "type": "string"
          },
          "lastPositionSeconds": {
            "type": "number"
          },
          "version": {
            "type": "integer"
          },
          "watchedCoverage": {
            "type": "number"
          },
          "watchedRanges": {
            "items": {
              "$ref": "#/components/schemas/WatchedRangeResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "courseStepId",
          "watchedRanges",
          "lastPositionSeconds",
          "watchedCoverage",
          "bestQuizScore",
          "completedAt",
          "version"
        ],
        "type": "object"
      },
      "LearningProposal": {
        "properties": {
          "agentRunId": {
            "format": "uuid",
            "type": "string"
          },
          "approvedCourseId": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "approvedCourseVersion": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "candidate": {
            "properties": {
              "channelName": {
                "type": "string"
              },
              "reason": {
                "type": "string"
              },
              "thumbnailUrl": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "videoUrl": {
                "format": "uri",
                "type": "string"
              }
            },
            "required": [
              "title",
              "videoUrl",
              "thumbnailUrl",
              "channelName",
              "reason"
            ],
            "type": "object"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ownerId": {
            "minimum": 1,
            "type": "integer"
          },
          "proposalVersion": {
            "minimum": 1,
            "type": "integer"
          },
          "state": {
            "enum": [
              "pending",
              "approved",
              "dismissed",
              "expired"
            ],
            "type": "string"
          },
          "videoSourceId": {
            "pattern": "^[1-9]\\d*$",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerId",
          "agentRunId",
          "videoSourceId",
          "proposalVersion",
          "state",
          "candidate",
          "expiresAt",
          "approvedCourseId",
          "approvedCourseVersion"
        ],
        "type": "object"
      },
...
      "QuizResponse": {
        "properties": {
          "courseStepId": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "maxAttempts": {
            "type": "integer"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            },
            "type": "array"
          },
          "schemaVersion": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "courseStepId",
          "version",
          "schemaVersion",
          "maxAttempts",
          "questions"
        ],
        "type": "object"
      },
...
      "RecordProgressDto": {
        "properties": {
          "endSeconds": {
            "minimum": 0,
            "type": "number"
          },
          "lastPositionSeconds": {
            "minimum": 0,
            "type": "number"
          },
          "occurredAt": {
            "type": "string"
          },
          "startSeconds": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "startSeconds",
          "endSeconds",
          "lastPositionSeconds",
          "occurredAt"
        ],
        "type": "object"
      },
...
      "RequestAdaptiveQuizDto": {
        "properties": {
          "endSeconds": {
            "minimum": 0,
            "type": "number"
          },
          "startSeconds": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "startSeconds",
          "endSeconds"
        ],
        "type": "object"
      },
...
      "SubmitAdaptiveQuizDto": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuizAnswerDto"
            },
            "maxItems": 5,
            "minItems": 5,
            "type": "array"
          }
        },
        "required": [
          "answers"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "enrollment": {
        "in": "cookie",
        "name": "studytube_enrollment",
        "type": "apiKey"
      },
      "internalApiKey": {
        "in": "header",
        "name": "x-internal-api-key",
        "type": "apiKey"
      },
      "mcpService": {
        "bearerFormat": "MCP service assertion",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "in": "cookie",
        "name": "studytube_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Authenticated StudyTube web and learning API",
    "title": "StudyTube API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/learning/agent-runs": {
      "post": {
        "operationId": "LearningController_createRun",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRunDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/agent-runs/{runId}": {
      "get": {
        "operationId": "LearningController_getRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/agent-runs/{runId}/approve": {
      "post": {
        "operationId": "LearningController_approveRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/agent-runs/{runId}/cancel": {
      "post": {
        "operationId": "LearningController_cancelRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/agent-runs/{runId}/next-learning-proposal": {
      "post": {
        "operationId": "LearningController_createNextLearningProposal",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProposal"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/agent-runs/{runId}/retry": {
      "post": {
        "operationId": "LearningController_retryRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
...
    "/learning/contexts/{contextId}/quiz-loops": {
      "post": {
        "operationId": "LearningController_requestAdaptiveQuiz",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d*$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAdaptiveQuizDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveQuizLoop"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/course-steps/{stepId}/progress": {
      "get": {
        "operationId": "LearningController_getProgress",
        "parameters": [
          {
            "in": "path",
            "name": "stepId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProgressResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      },
      "post": {
        "operationId": "LearningController_recordProgress",
        "parameters": [
          {
            "in": "path",
            "name": "stepId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d*$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordProgressDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProgressResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/course-steps/{stepId}/quiz": {
      "get": {
        "operationId": "LearningController_getQuiz",
        "parameters": [
          {
            "in": "path",
            "name": "stepId",
            "required": true,
            "schema": {
              "pattern": "^[1-9]\\d*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
...
    "/learning/proposals/{proposalId}": {
      "get": {
        "operationId": "LearningController_getLearningProposal",
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProposal"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/proposals/{proposalId}/dismiss": {
      "post": {
        "operationId": "LearningController_dismissLearningProposal",
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProposal"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/proposals/approve": {
      "post": {
        "operationId": "LearningController_approveLearningProposal",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveLearningProposalDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningProposal"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/quiz-loops/{quizLoopId}": {
      "get": {
        "operationId": "LearningController_getAdaptiveQuiz",
        "parameters": [
          {
            "in": "path",
            "name": "quizLoopId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveQuizLoop"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/quiz-loops/{quizLoopId}/submit": {
      "post": {
        "operationId": "LearningController_submitAdaptiveQuiz",
        "parameters": [
          {
            "in": "path",
            "name": "quizLoopId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAdaptiveQuizDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveQuizSubmission"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/learning/quizzes/{quizId}/attempts": {
      "get": {
        "operationId": "LearningController_listQuizAttempts",
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuizAttemptResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      },
      "post": {
        "operationId": "LearningController_submitQuiz",
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAttemptResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ],
  "servers": [],
...
}